pyproject.toml
[tool.poetry]
name = "edge-benchmarking-client"
version = "0.1.0"
description = "Client library for the AgrifoodTEF Edge Benchmarking Service"
authors = ["Andreas Schliebitz <a.schliebitz@hs-osnabrueck.de>"]
license = "Apache-2.0"
readme = "README.md"
packages = [{ include = "edge_benchmarking_client" }]

[tool.poetry.dependencies]
python = "^3.10"
requests = "2.32.3"
validators = "0.28.3"
python-dotenv = "1.0.1"

[tool.poetry.group.dev.dependencies]
black = "24.4.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
